requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Databricks and data access
databricks-sdk>=0.60.0
databricks-sql-connector>=3.0.0

# PostgreSQL async support (for Lakebase)
sqlalchemy>=2.0.0
asyncpg>=0.29.0
sqlmodel>=0.0.14

# ML and embeddings
numpy==1.26.3
pillow==10.2.0

# HTTP and utilities
httpx>=0.26.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Vector Search - FIXED VERSION
databricks-vectorsearch>=0.40

# HTTP client for async requests
aiohttp>=3.9.0